Stop the container when any process fails and silence logs by default (#5)
With these changes, the entire container will exit if either of the
zookeeper or kafka processes exits, rather than lingering in a silently
broken state.

While we're at it, we also silence the extremely verbose logs by
default, and instead simply log process lifecycle events.

If a process fails and we have to shut down, we do at least try to show
some helpful debugging logs.

The behavior can be changed by running the image with
`EXIT_ON_FAILURE=false`.